Started to implement Diagnostics System.
At a loss of what to do, I am starting at Diagnostics System, Different than Stride Engine they have it under core, but we will want to log everything sophisticatedly so user is never lost in the dust. So I have a dedicated a cs project to it to ensure we have grips on whole system.